post store issue fixed

+++ app/Services/PostService.php
@@ -3,24 +3,27 @@
 namespace App\Services;
 
 use App\Models\Post;
+use Carbon\Carbon;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Str;
 class PostService
 {
     public function createPost($request)
     {
-       $post = Post::create([
+        $post = DB::table('posts')->insert([
             'user_id' => Auth::user()->id,
             'uuid' => Str::uuid()->toString(),
             'total_views' => 1,
             'description' => $request->barta,
+            'created_at' => Carbon::now()
+
         ]);
-     
+
     }
 
     public function remove($id){
         $loggedInUser = Auth::user()->id;
-        $post = Post::where('user_id', $loggedInUser)->find($id);
-        $post->delete();
+        $post = DB::table('posts')->where('user_id', $loggedInUser)->where('id',$id)->delete();
     }
 }
